fixes #811
Fixed jarbled output for certain formal corrections in the formal correction header.
The js routines weren't factoring the possibility of html markup in the formal correction comment which happens by design as CRs are translated to <br/>s.
Now when a multi-paragraph comment is encountered, only the first paragraph is shown in the formal correction header.  
(The entire comment is still shown in the associated annotation pop-up dialog of course).

git-svn-id: http://svn.topazproject.org/svn/branches/0.8.2.2@4963 48088309-b40f-0410-b149-b14953253675