Refactor Mapper interface to reflect the fact that not all properties that OTM manages have a direct RDF persistence. So Mapper is the super interface that only contain a few properties common to all Mappers. The old Mapper is renamed as RdfMapper. Additional Mapper sub-classes are Id, Blob, Embedded and OQL projections in a view.
Also ClassMetadata now holds a list of Embedded Mappers. This made 
the OQL hacks via the POJO binders to get to the Embedded class disappear. 

That leaves the View projection binding the only place that is POJO
aware. (There is a bit of a chicken-and-egg situation here.)
(See r4215, r4590, r4741, r4850 for the changes related to supporting
non POJO bindings)

This also addresses #818. Embedded classes still are not treated as 
a form of multiple-inheritence yet.




git-svn-id: http://svn.topazproject.org/svn/head@5141 48088309-b40f-0410-b149-b14953253675