Addresses #525, #526, #527 and #855. Here are the list of things done here:
 * Removal of org.plos and org.plosone from <groupId> and changed that to
   org.topazproject
 * Removal of plosone and plos strings from <artifactId> and changed that to
   ambra 
 * Removed topaz string from <artifactId> of ambra packages
 * Made ambra.xml the default configuration file

Please note: Ambra is already broken in 'head', and because of really slow
network speeds I have not been able to test the rpms completely. I am going to
have to do a clean build on black to test this thoroughly. If you don't want
this to interfere with your work, do not do an 'svn update'.


git-svn-id: http://svn.topazproject.org/svn/head@5412 48088309-b40f-0410-b149-b14953253675