Make Jotm instance a singleton.
While multiple Jotm instances can be created, they use an instance of Current
underneath which in turn is effectively a singleton because most of the fields
are static. Since we basically always create identical Jotm instance this
shouldn't really be a problem, except that it may lead to exceptions on cleanup.


git-svn-id: http://svn.topazproject.org/svn/head@5435 48088309-b40f-0410-b149-b14953253675