Convert ObjectInfo representations to use the Blob support available from OTM. This required a bit of data-migration (to get rid of the ugly format-contentType and format-objectSize dynamic predicates and to have topaz:hasRepresentation point to the Representation object as opposed to a literal value).
To have minimum impact on the ongoing development effort, this migration is done
automatically at stratup by ambra. It can be used in production also, but if
required can be turned off by removing this listener from web.xml.

Also note that there is no change done to ingest. Usage of otm-models and
the blob will be done as part of #439. Till then the data-migration is done 
automatically as a post-processing of ingest.

This addresses #776 and #856. (RepresentationInfo is now gone)



git-svn-id: http://svn.topazproject.org/svn/head@5475 48088309-b40f-0410-b149-b14953253675