Search results are shared, but not concurrently. Protect it with a lock.
Also added a mode for degrading to non-cached instead of waiting 
forever. It is probably futile since the failure to acquire the lock is 
likely an indication of a dead-lock somewhere - so may be reporting
an error is better here rather than the degraded mode. Needs more
testing to determine that.



git-svn-id: http://svn.topazproject.org/svn/head@5521 48088309-b40f-0410-b149-b14953253675