Another renaming:
 * Renamed fedora to topaz-fedora
 * Renamed topaz-tools to ambra-tools
 * Renamed cas to ambra-cas

I have also changed the directory names to correspond to artifactId names to
conform to Maven convention as pointed out by Pradeep. I have not tested this
rpms on black as upload of rpms to black is taking a loooong time. I will build
and test on black and let you know if everything went okay.



git-svn-id: http://svn.topazproject.org/svn/head@5539 48088309-b40f-0410-b149-b14953253675