addresses #895 Dojo upgrade (0.4.3 -> 1.1) - INITIAL COMMIT.
Not everything is fully functional yet but this changeset represents the lion's share of work for this upgrade.
Of note:
1) Abandoned fine grained "dojoRequire_{xxx}.ftl" include files opting simply for 2 versions:
    a) dojoRequire_default.ftl - This is a minimal set of js dependencies targeted for the home page and non-functional related pages.
    b) dojoRequire_full.ftl - This is the full set of js dependencies targeted for the article pages and adjacent supporting pages (annotation, comment, rating editing, slideshow, et al.).
The ultimate plan here is to create a js resource deploy script that compresses all js dependencies into a single compressed file to be used for production.  Dojo 1.1 has a tool for this.

2) All core api changes have been propagated.  The majority of the work yet to be completed lay in our custom dojo widgets namely:
   a) annotation_add.ftl
   b) commentDialog.ftl
   c) loadingCycle.ftl
   d) ratingDialog.ftl

The site is navigable and articles are viewable.  Annotations, Comments, Ratings are currently only viewable.  No editing functionality for these article decorations currently works.  My goal to to have this upgrade completed by end of next week (5/9/08).


git-svn-id: http://svn.topazproject.org/svn/head@5581 48088309-b40f-0410-b149-b14953253675