Update user-service to not trip on the changes in the 0.8.3 version of OTM. OTM is now state based. So this means once an object is attached to the session, OTM will keep track of any state changes and save it automatically.
PlosOneUser was updating the UserProfile object with no intention of saving
it. Therefore everytime we create a PlosOneUser, we detach the UserProfile
from Otm Session so that OTM will not save it.

The other is related to a change in the default cascade. It used to be
all + deleteOrphan. Now it is just 'all' by default. So update the
user model to explicitly set the cascade mode.

Addresses #910. So create/update user/profile and login/logout
should all work now.



git-svn-id: http://svn.topazproject.org/svn/head@5585 48088309-b40f-0410-b149-b14953253675