More formatting fixes.
Folks, please read the code formatting spec on our wiki. It '''really''' is not
that hard. If you don't like something there, feel free to bring it up for
discussion, but keep in mind that consistency is crucial as new developers come
on board and start to adopt the code.



git-svn-id: http://svn.topazproject.org/svn/head@5632 48088309-b40f-0410-b149-b14953253675