Make 'edit-preferences' work again. Addresses #910.
Turns out there is one more nuance to that 'evict' saga for the UserProfile 
object. A manual save needs to be called since the UserPreferences may have
been previously evicted. Also after saving, the UserProfile object, needs to
be evicted again since it is likely to be from the PlosOneUser object stored
in the Http-Session.

There could still be a potential problem with this evict strategy.
Associations that reference the evicted object can 're-attach' it. So in 
short it is not a good idea to muck with a persistent object unless the goal
is to change its state in the database. So the PlosOneUser object needs to be
fixed for this reason - or use a different Session object to load it.



git-svn-id: http://svn.topazproject.org/svn/head@5633 48088309-b40f-0410-b149-b14953253675