Disable the 'addSelf' implicit smartcollection rule added to Journal context OTM filters. This rule limits Journal objects to be loaded only from within a Journal's context. This creates trouble with features that require to see other journal instances such as 'publish' and the reverse map from an Article to the list of Journals it belongs to.
The reason why this used to work before was because the cache used to
maintain a global/shared Journal object instance in cache. Now that we
have moved to "hydrated" instances from the second level cache in OTM,
get-filtering comes into play and makes a Journal instance visible only
in its context.



git-svn-id: http://svn.topazproject.org/svn/head@5718 48088309-b40f-0410-b149-b14953253675