Cleanup with regards to files being created in 'ingested' queue for error conditions.
With regards to ingestion:

 * Pulled out creation of Xref and move to ingested queue outside of txn.

With regards to delete:

 * Made the transaction scope around service instead of action
 * Pulled out deletion of xref and move back to ingestion queue outside of txn.

Please note: this is not a very clean fix as errors related to IO will not cause
a transaction roll back and will need to be fixed manually. But given the common
cases this is the best fix as of now.

Ronald, Pradeep: I have tested this, but would appreciate you guys doing a code review as usual.



git-svn-id: http://svn.topazproject.org/svn/head@5907 48088309-b40f-0410-b149-b14953253675