revert attempt to override user template .ftl files on a per-journal basis.
defining a couple of variables in user_shared_template.ftl instead to provide what we need for 0.9.

ideally, these strings would be pulled from config, and jon has agreed to take a look at this after this commit.

finally, this commit introduces an error when moving from the alerts tab to the preferences tab, as the variables defined in user_shared_template.ftl are not available. hopefully this will be fixed by moving this stuff to config, although it's a bad solution for the 'privacy statement' variable. will discuss with jon - perhaps we should keep things static and plos-focused until after the push to production pre-release in this case.

git-svn-id: http://svn.topazproject.org/svn/head@5931 48088309-b40f-0410-b149-b14953253675