Improve fedora ingest and delete performance. Fedora calls are not cheap, so the changes here attempt to reduce the number of calls by sometimes guessing at the operation to execute and then examining the error if necessary to find out what to do instead.
Specifically, ingest now directly tries to create an object, add a datastream,
or modify a datastream, instead of first asking fedora whether the object or
datastream already exists; if an error occurs, that error is analyzed to
figure out which operation to try instead. Furthermore, there is a new method
getFirstIngestOp() which can be overriden by subclasses to help guess the
first operation to try; for example, blobs that store only a single datastream
per object and that are never (or hardly ever) updated can specify that an
ingest should be tried first.

For delete, canPurgeObject() has been changed to be invoked before the first
fedora call; this reduces the max number of calls from 3 to 2, and in the case
where there is always only one datastream per object (such as for annotations)
that implementation can return 'true' to further reduce the calls down to 1.

The main drawback of this approach is that it relies on observed behaviour of
a given fedora version (2.1.1 at the time of this writing), since the error
behaviour of the various API-M calls is not defined.

The results of these optimizations for pone.0002145 are reduction in ingest
time from 55s to 45s, and a reduction in delte time from 47s to 37s.

Further speedups for bulk adds and deletes (such as ingesting/deleting an
article) could be achieved by caching information about the objects and their
datastreams, probably on a transaction basis (i.e. clear the cache at the end
of the transaction).


git-svn-id: http://svn.topazproject.org/svn/head@5968 48088309-b40f-0410-b149-b14953253675