fixes #962, fixes #912
Fixed unwanted page scrolling for:
 - closing the annotation dialog
 - clicking on existing annotation link
 - clicking on a formal correction link

Reverted the fade in/out time for the annotation dialog from 1/2s to a 1/4s like it was pre dojo1.1.

Opted to NOT show the "loading cycler" widget when an xhr call is made for retrieving an annotation from server as this was interfering with scrolling.  We may want to confirm this from a usability point of view as this may cause confusion if there is excessing network latency as there will be no indication, save for the grayed background, that a server call is in process.

git-svn-id: http://svn.topazproject.org/svn/head@5975 48088309-b40f-0410-b149-b14953253675