Enable gzip compression for text files. (Not quite true: the filter is also applied to *.action which includes fetchObject that loads jpeg files from Fedora)
The filter is from ehcache. So no new jars and also it seems to be 
doing a pretty good job of handling the edge cases.

Addresses #518.

Note that I can't observe any difference in response times.
However this does do a nice job of compressing .js and
the article pages. So folks with low bandwidth connections
would benefit. Take a look at it with 'yslow' to see.
(Clear your browser cache if something doesn't show up as
gzipped but it is supposed to be.)

Since I couldn't measure this and I am going by 'best practices'
recommendations, if anyone feels that this has any negative effect,
we can disable this filter any time.



git-svn-id: http://svn.topazproject.org/svn/head@6000 48088309-b40f-0410-b149-b14953253675