Must do a flush() before looking at the annotation list from the cache. This is now evidence that we should really be moving the query cache management also into OTM. (or quasi-OTM like the Object-Cache)
This was causing the delete-flag on an annotation
from only deleting the flag annotation but not changing
the flag state of annotations. Results in the puzzling
log statements as reported in #970.

This closes #970. The flagged state of some annotations
will remain. It is harmless (as far as I can see). But
if you don't like to see those messages, data needs to
scrubbed. ie. Reset the 'flag' bit from the state of
all annotations that are no longer flagged. See 
org.plos.Constants for the value of the flag bit.



git-svn-id: http://svn.topazproject.org/svn/head@6041 48088309-b40f-0410-b149-b14953253675