To deal with the never ending stream of problems related to configuration changes. Execute:
{{{
/usr/local/topaz/bin/rungroovy /usr/local/topaz/scripts/ambraConfiguration.groovy
}}}

to dump the merged configuration to screen. Please note that this is an example
script and not a supported tool. Feel free to change if you want.

The idea is that you can take and save a dump before RPM upgrade and compare it
with a dump after upgrade to figure out the changes. I would recommend passing
it through another utility like '''sort''' to make it easier.



git-svn-id: http://svn.topazproject.org/svn/head@6075 48088309-b40f-0410-b149-b14953253675