Minor cleanup and a bug fix on the flagged Reply look-up.
Note that Reply does not have the PUBLIC_MASK set.
So either those all should have the PUBLIC_MASK added
or we eliminate this bit all together for both Annotations
and Replies. 

There is also a bit of an inconsistency with regards to the 
DELETE_MASK. It is only being used for Replies and not for Annotations.

Not sure what the reasons are for these bit settings. The only
one that makes sense is FLAG_MASK (really should be renamed to FLAG_BIT).
So once we decide on this, we can do the data migrations and make
things consistent.

See #857 for related issues. 



git-svn-id: http://svn.topazproject.org/svn/head@6080 48088309-b40f-0410-b149-b14953253675