Optimized the representation migration query. Now it only takes about 4 seconds for the Migration to determine there is nothing to do (instead of the 2 mins. This is when there are a 163463 representations)
No change in the first run. It takes about 15 - 20 minutes.
(This is when closer to a 800,000 statements are migrated.
  Of which about 500,000 are for migrating the Representations
  These all belong to 4423 articles)

Make sure to run with a large timeout value for txns for the 
first time Ambra is run on the 0.8.2 data.
(
eg. -Dbitronix.tm.timer.defaultTransactionTimeout=3000
   default of 10mins (ie. '600') is not enough.
)

Note: all numbers above are for moody.topazproject.org with
      a SingleBoxInstall. Mulgara is restored from a backup
      of production mirror - version 0.8.2  



git-svn-id: http://svn.topazproject.org/svn/head@6123 48088309-b40f-0410-b149-b14953253675