This change fixes Safari js issues reported in: http://joborder.plos.org/view.php?id=6737#bugnotes.
Basically, dojo's console.debug statements don't work in dojo's non-debug mode in Safari 3.1.1.  This is a known issue and in fact the issue has since been fixed in the nightly builds for both dojo and Safari:

http://trac.dojotoolkit.org/ticket/6849
https://bugzilla.osafoundation.org/show_bug.cgi?id=12146

This change is temporary until we upgrade to the next version of dojo which will surely contain this fix.

git-svn-id: http://svn.topazproject.org/svn/head@6159 48088309-b40f-0410-b149-b14953253675