reduce maxProcessors and maxThreads on AJP container.
typically mulgara and ambra get into a hang when there are so many threads waiting for a mulgara lock that, by the time mulgara is free the waiting process has timed out. i'm hoping that reducing the max number of connections will keep this from happening - instead users will get a 503 error earlier in an overload situtation.

i would set this to 50 based on what i've seen in 0.8 production, but suspect 0.9 will handle more concurrent connections.

git-svn-id: http://svn.topazproject.org/svn/head@6180 48088309-b40f-0410-b149-b14953253675