Fix the issue where the author list was not showing up in the browse pages after an ingest and publish.
UserProfileInfo and CitationInfo are "usage-model" entities
and therefore they should have their properties as a reference
to the "content-model" entities FoafPerson and Citation. Since
currently there is no way to configure the Refernce Definition
in OTM, a local alias definition was added. Revert that
portion when this becomes available in OTM.

Also missing properties for an entity must trigger a load
from the store - rather than an empty value list. This would happen
if for the same Id there are two entities in the cache. 
(As in this case - CitationInfo and Citation).



git-svn-id: http://svn.topazproject.org/svn/head@6189 48088309-b40f-0410-b149-b14953253675