Annotations and Replies service had an old assumption about permissions service being on a different transaction. So it was "handling" the exceptions from permissions service and tries to "undo" the actions prior to that.
Now that permissions service is part of the same transaction and 
the @Transactional annotation on the permission service methods will 
mark the transaction as 'read-only' and making the whole 'undo' operation
useless. 

So removed all that additional code and simplified the logic.

This also fixes #857. deleteReplies method was passing a bogus URI
to the permissions service causing it to throw an exception and
marking the transaction as roll-back. ReplyService caught the exception
and logged it at DEBUG and then proceeded to do the 'undo' and much
later when the transaction was committed we see 'the attempt to use read-only 
transaction for writes' exception (as reported in the ticket).

So no updates were done. However an interesting side effect occured.
The ArticleAnnotationCache that caches the list of annotations for a
target, was listing '0' annotations for the Reply. So a subsequent
visit to the adminTop.action page will not show the Reply as flagged.

The culprit there really is that we made a conscious decision to have
cache population action as non-transactional. (Only cache updates are
transactional now). This is to avoid multiple concurrent transactions
from trying to load the same missing entry. (See Cache.SynchronizedLookup)

This is only an issue in the case of a roll-back. (As in #857). However
the heavily used Object-Caches do not use non-transactional populates and 
doesn't seem to have that much of an impact anyway. So I am wondering
if we should try that approach for all caches - and make the roll-back
behavior correct. 



git-svn-id: http://svn.topazproject.org/svn/head@6295 48088309-b40f-0410-b149-b14953253675