removed deprecated config_home.js from journal templates
added init_home.js overrides and configured correctly for various journal tab sets.

note that path in default init_home.js to eg. /article/mostCommented.action is incorrect - struts is configured for /mostCommented.action even though the ftl resides in the article folder. this URL works in the default context through some kind of cascade, but does not work in the overlay journals, with the result that the default tab content .ftls were getting displayed instead of journal-specific tab content. this commit fixes this bug.

i'm not sure if it's worth fixing the URLs in the default init_home.js for consistency's sake or not so i'm leaving it for now.

git-svn-id: http://svn.topazproject.org/svn/head@6309 48088309-b40f-0410-b149-b14953253675