reduce template_update_delay to 600s, which makes a little more sense for a production site. the previous value of 600000s (about one week) is too long to wait for static content updates.
for plos, frequently used templates are accessed many times in a 10 minute period (70 on average for the templates involved in the main fetchArticle.action) so we still get most of the benefits of a longer delay on these templates. less frequently used templates suffer more, which is reasonable IMO given the trade-off.

git-svn-id: http://svn.topazproject.org/svn/head@6310 48088309-b40f-0410-b149-b14953253675