Move OTM annotations to getter and setter Methods. The recommended convention is to annotate the setters.
An additional restriction is that all properties
that are to be persisted, must have an OTM annotation on
its setter (or getter but not both). This closes #886.
(Note the side effect mentioned on #886.)

Another side effect is that @Embeddable is no longer
required and therefore is eliminated.

No change in @UriPrefix. This allows annotating setters
(or getters) with @Predicate annotation without specifying
any predicate-uri. The predicate-uri is computed from
the @UriPrefix as before - but an @Predicate annotation
must exist on the setter (or getter).

Also note that both setter and getter must exist for
Entity properties. For Views, only setters are mandatory.


git-svn-id: http://svn.topazproject.org/svn/head@6374 48088309-b40f-0410-b149-b14953253675