Added reference to @Predicate annotation. Closes #989.
Note that there is a bit of ugliness in 'tri-stating'
the boolean attributes in @Predicate annotation. If
this is too much of a pain, override of reference
can be excluded for boolean attributes.

Also removed the hack in #6189, and introduced another
(*sigh*) for rdfs:range comparison. This will only
be fixed after the 2-pass configuration support in
SessionFactory to resolve forward references.


git-svn-id: http://svn.topazproject.org/svn/head@6410 48088309-b40f-0410-b149-b14953253675