Support for java generics. No runtime dynamically defined entities - but support for statically defined subtypes of generic base classes. eg.
{{{
@Entity
class Foo<T> {
  T getFoo() {...}
  @Predicate()
  void setFoo(T v) {...}
}

@Entity
class Bar extends Foo<Integer> {
}

Bar b = new Bar();
Foo f = new Foo<Double>();
}}}

'b' will be persisted correctly. But 'f' will not since
by default getFoo() is supposed to be an rdfs object-property
and therefore the insert will fail since a URI is what
is expected for the 'foo' property.

Note also that java.util.Collection of generics and arrays
of generics are also supported. eg. List<T> getFoos() or
T[] getFoos() both are valid generic base class propery types.

T can resolve to any association or literal. However
Foo<T> as a return type for getFoo() is not supported.
From the example above, the resolved Foo<Integer> is not
recognized as Bar in this case. So the generic type variable
'T' will be ignored in this case.

This closes #780.


git-svn-id: http://svn.topazproject.org/svn/head@6432 48088309-b40f-0410-b149-b14953253675