FetchType.lazy now applies to properties as opposed to how the individual associated instances are loaded. The ItqlStore is still doing 's p o' queries. So large CollectionType.PREDICATE collections at least have one statement loaded per row. So if this represents a literal field, the field is effectively 'eager' loaded and if this represents an association the id is eager loaded.
Tests and such that assumed eager loaded properties are modified to
set the fetch type to eager. And so does the assumption in
ambra that prompted r5583 (hence the update to 'eager' for the back-pointer
to Article from ObjectInfo.)

This reverts the special case handling of java.util.Set in r4605 to
address #627. r4605 was delaying the creation of association
instances specifically for java.util.Set. Now it is done for
all associations and for all rdf:list and rdf:bag fields.


git-svn-id: http://svn.topazproject.org/svn/head@6486 48088309-b40f-0410-b149-b14953253675