Add support for multiple super entities (multiple-inheritance). In Java binding this means support for interfaces.
Java interface support:
  * Every interface that OTM annotation parser encounters is an entity
  * Annotations on a get/set method from an interface is sufficient to
    define a persistable property for any of the implementing classes.
  * Generics support is extended to interfaces and therefore a
    parameterized implementation can supply the concrete serializable
    types (or associations as the case may be)
  * A class that defines no rdf:type of its own will have its rdf:type
    set to the union of rdf:type values of its superClass and all of
    the interfaces that it implements. In OWL terms the class definition
    is the sub-class of the intersection of all of the individual classes.
    (TODO: OWLGenerator needs to be updated)

Note that for using interfaces in associations etc., the same rules of
what a loadable entity definition is etc. holds. ie. There has to be a
model/named-graph defined in the entity definition for OTM to do the
query on.

Also note that OTM does not do dynamic entity definition creation based on
the combination of rdf:type it finds on a load. So all instances it creates
are based on a concrete instance that the application has defined.

This closes #1023.


git-svn-id: http://svn.topazproject.org/svn/head@6526 48088309-b40f-0410-b149-b14953253675