Fix up permissions for annotations and replies. Note that listAllReplies() will prune the tree if the caller does not have permission to view a node.
'replies:getReplyInfo' permission is granted as part
of making a reply public. However currently this is
permitted by permitted-anonymous-operations xacml policy
rule (assuming default xacml rules).

The existing replies in the database does not have those
grants and therefore require the permitted-anonymous-operations
rule to permit this. To make things consistent, the
'replies:getReplyInfo' permission needs to be granted
in the grants named-graph. Only then it is safe to
remove the rule from permitted-anonymous-operations.

Note also that 'replies:deleteReply' permission is
revoked as part of making the reply 'public'. This
means with the default ambra policies only an admin
will be able to delete the reply. The 'creator' can
not delete it.

The existing replies in the database does not have those
revokes and therefore can be deleted by the 'creator'.
To make things consistent, the replies:deleteReply permission
needs to be revoked in the revokes named-graph.

This closes #441.


git-svn-id: http://svn.topazproject.org/svn/head@6606 48088309-b40f-0410-b149-b14953253675