Basic full-text search support in OTM.
To enable full-text indexing of a field add the @Searchable annotation to that
field (works for blobs as well as predicate fields). To search using the
full-text index(es) OQL now provides a new function search(); Criteria support
will not be implemented due to Criteria being deprecated.

This implementation, while functional, is far from perfect. It assumes heavily
that the triple is mulgara with the lucene-resolver, with only some minimal,
incomplete abstractions for general search stores. Also, a number of attributes
for the Searchable annotation and SearchableDefinition are not implemented yet,
such as 'tokenize', 'analyzer', and 'boost' (these are pending implementation
in the mulgara lucene-resolver).

git-svn-id: http://svn.topazproject.org/svn/head@6667 48088309-b40f-0410-b149-b14953253675