Moved createGraph and dropGraph from TripleStore to Session. All references to these methods have now been updated to use a Session, creating one if it didn't exist before, and establishing a transaction for writing to.
A number of tests attempt to drop graphs to ensure a clean slate, and catch the
exception if the graph did not exist. However, an exception like this will now
invalidate the writing transaction. This means that tests doing
dropGraph/createGraph operations need to start and stop transactions in a whole
lot of boilerplate code. To make this easier, the convenience methods
createGraphInTx() and dropGraphInTx() have been added for use in these tests
(they should not be used in normal code). If no transaction exists when these
methods are called, then they create a new transaction for the create or drop
operations, and commit this update before returning. If a transaction already
existed, then these methods operate the same as the normal createGraph and
dropGraph methods.


git-svn-id: http://svn.topazproject.org/svn/head@6678 48088309-b40f-0410-b149-b14953253675