Fixes #1053
Replies (also known as "responses") associated to an Annotation are now preserved when that Annotation's "type" (Minor Correction, Retraction, Formal Correction, or Note) is changed.  Previously, a new Annotation (with a new DOI) was created and the "replies" from the old Annotation were modified to refer to the new Annotation.  Now, the new Annotation uses the same DOI as the old Annotation and the "replies", permissions, etc (from the old annotation) are not changed.

Also, there was a bug in {{{adminTop.ftl}}} which was deleting the Annotation when the Annotation type was changed to "Note". This bug (hidden by the previous Annotation-type-changing functionality) manifested to the administrative user merely as the "flag" not being automatically deleted when an Annotation had its type changed to "Note". It was an annoyance, but nothing more, which is probably why the bug was left in place. This issue has also been corrected.

fixes #1053

git-svn-id: http://svn.topazproject.org/svn/head@6726 48088309-b40f-0410-b149-b14953253675