Added support for Streaming Blobs. Fedora and SimpleBlobStore has been updated to keep txn data in temporary files instead of in memory as before. This now allows streaming of large Blobs thru OTM. In addition the OTM blob support has been revamped to provide both 'managed' and 'unmanaged/copy' Blob fields.
Managed Blob fields must be used for large blobs. Unmanaged
ones are for simplicity where OTM copies the contents of the
blobs directly into the fields and does state tracking and
writes out any changes back to the BlobStore.

There is a bit of restriction on how Managed Blob fields can
be used. Since OTM manages these fields, application does not have
any control over when/how these fields are created. OTM creates
them on a load from store or on a saveOrUpdate(). ie. OTM is
the 'factory' class for these fields. Once created by OTM, the
application can read/write to the data-streams of these blobs
or even delete the blobs.

Because OTM is the factory for Managed Blobs, it can only create
the following types of objects:
  * org.topazproject.otm.Blob
  * javax.activation.DataSource
  * java.io.InputStream

Others may be added later. Note that the java.io.InputStream
essentially provides a read-only access to the Blob.

For Unmanaged/Copy Blob fields, an additional support has been
added for fields that are Serializable. This makes it possible
to define things like: '@Blob String getText();'. The Strings
are expected to be in UTF-8 encoding. (Other encodings may be added
later by modifying the @Blob annotation)

So far no change to the config to support this. The restriction
of byte[] on @Blob fields is now lifted.

Addresses #946.


git-svn-id: http://svn.topazproject.org/svn/head@6733 48088309-b40f-0410-b149-b14953253675