Switch Ambra to use the new OTM full-text search support. On the front-end side things still use and build lucene queries; these are then parsed and translated into an OQL query. This way nothing really changes from the users point of view.
The search-server is now no longer necessary and has been removed; and bringing
up ambra with jetty and embedded mulgara/blob-store now supports searching too.
The configuration for using mulgara's distributed resolver to split mulgara
into two instances, with and without the lucene queries and storage, has not
been done yet.

A couple things don't work quite yet either: there are some bugs with the -foo
(i.e. the "without words") translation in complex cases, and the boost factors
are not applied. Most everything else should work, though.

The hits are evaluated lazily (i.e. the loading of the article info is done
only when those hits are viewed). Also, this is done in a way such that
multiple clients running the same query (most likely a spider) will queue up
as necessary and get those results as soon as they are ready without resorting
to re-running the query if things time out.

Regarding caching, the query results are cached but the cache is never
explicitly invalidated; instead it is assumed the ttl on the cache (currently
configured to 10min) will be sufficient. The problem is it's basically
impossible to figure out when an article is published whether it would show up
in the search without re-running the query.

Lastly, because the query is filtered just like other OQL queries, the returned
results are now already limited to the journal, so the total-results estimate
is now more likely to be close to the real value (unpublished articles will
still cause the results to be further filtered on demand, though).

Addresses #744.

git-svn-id: http://svn.topazproject.org/svn/head@6771 48088309-b40f-0410-b149-b14953253675