Fixes #867
Administrative pages for managing the review of Annotations and Annotation Replies now both contain links to the Article and/or Annotation in question.  Links which were created with the {{{../}}} methodology now use {{{freemarker_config.context}}} for all URLs.  Creator name is now used for the text in the link to the creator's profile.  Page formatting changed to improve readablity.

Also completes Job Order 5330

git-svn-id: http://svn.topazproject.org/svn/head@6781 48088309-b40f-0410-b149-b14953253675