Rework Fedora BlobStore implementation to use Fedora's upload scratch area as the txn workspace. It is no longer a FileBackedBlobStore implementation and does *not* require a local tmp directory and avoids an expensive copy during ingest.
This will *not* work on the 'stock' 2.1.1 Fedora. It
requires a patched UploadServlet on Fedora that allows
download of the uploaded content.

Note also that the fedora-client is now using chunked transfer
encoding to upload - this also avoids the expensive Content-Length
computation that the commons-http-client library was performing.

Both these changes together has cut down the ingest times in half
for large ingests. So the txn timeout for ingest has been set to
30 minutes in the applicationContext.xml.


git-svn-id: http://svn.topazproject.org/svn/head@6897 48088309-b40f-0410-b149-b14953253675